requirements.txt
Flask==3.0.3
tensorflow==2.15.0
Pillow==10.0.1
numpy==1.24.3
opencv-python-headless==4.8.1.78
gunicorn==21.2.0
Werkzeug==3.0.3